+++ spring-general-lab/src/main/java/com/gusrubin/lab/springgenerallab/infrastructure/config/ApplicationConfig.java
@@ -8,6 +8,7 @@ import org.modelmapper.convention.MatchingStrategies;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.data.web.config.EnableSpringDataWebSupport;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 /**
@@ -16,6 +17,7 @@ import org.springframework.scheduling.annotation.EnableScheduling;
  */
 @Configuration
 @EnableScheduling
+//@EnableSpringDataWebSupport
 @EnableJpaRepositories(basePackages = "com.gusrubin.lab.springgenerallab.infrastructure.database")
 public class ApplicationConfig {