+++ spring-general-lab/src/main/java/com/gusrubin/lab/springgenerallab/infrastructure/database/entities/PageableResourceEntity.java
@@ -1,0 +1,33 @@
+package com.gusrubin.lab.springgenerallab.infrastructure.database.entities;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ * @author Gustavo Rubin
+ *
+ */
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "pageable_resource")
+public class PageableResourceEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(name = "text")
+    private String text;
+
+}