+++ spring-general-lab/CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Runtime scheduling example
 - Freezing protection example
+- Pageable example